serving/webservice/requirements.txt
numpy~=1.19.2
tensorflow~=2.5.0
pillow~=8.0.1
requests~=2.24.0
python-multipart~=0.0.5